requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clear==2.0.0
cloudinary==1.44.0
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-minio-storage==0.5.8
django-tailwind==3.8.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==3.0.2
minio==7.2.15
msgpack==1.1.0
numpy==2.1.1
oauthlib==3.2.2
onionsearch==1.3
packaging==24.2
pandas==2.2.2
Pillow==10.0.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.5.1
scipy==1.14.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.3
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.13.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.9.0
